+++ src/main/java/poetry/internal/database/native/CursorItems.kt
@@ -1,0 +1,5 @@
+package poetry.internal.database.native
+
+import android.database.Cursor
+
+fun Cursor.hasItems() = count != 0