+++ src/main/java/poetry/internal/database/native/CursorUse.kt
@@ -1,0 +1,9 @@
+package poetry.internal.database.native
+
+import android.database.Cursor
+
+internal fun <Result> Cursor.use(block: () -> (Result)): Result {
+	val result = block()
+	close()
+	return result
+}