+++ src/main/java/poetry/internal/reflection/Annotations.kt
@@ -1,0 +1,3 @@
+package poetry.internal.reflection
+
+internal typealias AnnotationMap = HashMap<Class<out Annotation>, Annotation?>